requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.4
cffi==1.11.5
cryptography==2.2.2
idna==2.6
multitasking==0.0.7
paramiko==2.4.1
pyasn1==0.4.2
pycparser==2.18
PyNaCl==1.2.1
PyYAML==3.12
scp==0.10.2
six==1.11.0
systemd==0.16.1
virtualenv==15.2.0
